Get forms by user permission

+++ routes/dynamic.js
@@ -3,6 +3,9 @@ const { default: mongoose, Schema } = require("mongoose");
 const router = express.Router();
 const multer = require("multer");
 const fs = require("fs");
+const { auth, verifyRootLevel } = require("../middleware/auth");
+const jsonwebtoken = require("jsonwebtoken");
+const AccountModel = require("../models/AccountModel");
 
 const storage = multer.diskStorage({
   destination: function (req, file, callback) {
@@ -22,7 +25,7 @@ const storage = multer.diskStorage({
 const upload = multer({ storage: storage });
 
 // Creates a form by requested fields
-router.post("/create-form", upload.single("file"), (req, res) => {
+router.post("/create-form", verifyRootLevel, upload.single("file"), (req, res) => {
   try {
     const formStructure = JSON.parse(req.body.formStructure);
 
@@ -55,72 +58,123 @@ router.post("/create-form", upload.single("file"), (req, res) => {
 });
 
 // Returns list of forms
-router.get("/get-forms", (req, res) => {
+router.get("/get-forms", auth, (req, res) => {
   try {
-    if (!mongoose.models.formSchemas) {
-      createSchemasModel();
+    const decoded = req.account;
+    if (decoded.role === "admin") {
+      if (!mongoose.models.formSchemas) {
+        createSchemasModel();
+      }
+      AccountModel.findById(decoded.account_id, { "allowedForms.formId": 1, _id: 0 }).then(
+        (formResponse) => {
+          let formIds = [];
+          Object.values(formResponse.allowedForms).forEach((value) => {
+            formIds.push(value.formId);
+          });
+          mongoose.models.formSchemas
+            .find({ _id: { $in: formIds } })
+            .then((response) => {
+              res.json(response);
+            })
+            .catch((err) => {
+              console.log;
+              res.status(404).send();
+            });
+        }
+      );
+    } else if (decoded.role === "root") {
+      if (!mongoose.models.formSchemas) {
+        createSchemasModel();
+      }
+      mongoose.models.formSchemas
+        .find()
+        .then((response) => {
+          res.json(response);
+        })
+        .catch((err) => {
+          console.log;
+          res.status(404).send();
+        });
     }
-    mongoose.models.formSchemas
-      .find()
-      .then((response) => {
-        res.json(response);
-      })
-      .catch((err) => {
-        console.log;
-        res.status(404).send();
-      });
   } catch (error) {}
 });
 
-router.put("/update-form", upload.single("file"), (req, res) => {
+router.put("/update-form", auth, upload.single("file"), async (req, res) => {
   try {
     const formStructure = JSON.parse(req.body.formStructure);
-    if (!mongoose.models.formSchemas) {
-      createSchemasModel();
-    }
-    const formSetupModel = mongoose.models.formSchemas;
+    const isAllowed = await checkPermission(req.account, formStructure[2].form_id);
+    if (req.account.role === "root" || (req.account.role === "admin" && isAllowed)) {
+      if (!mongoose.models.formSchemas) {
+        createSchemasModel();
+      }
+      const formSetupModel = mongoose.models.formSchemas;
 
-    const formSetupData = {
-      formName: formStructure[1].formName,
-      description: formStructure[1].description,
-      formDetails: formStructure[0],
-      icon: `/img/formIcons/${req.file.filename}`,
-      primaryColor: formStructure[1].primaryColor,
-    };
-    formSetupModel
-      .findByIdAndUpdate({ _id: formStructure[2].form_id }, formSetupData)
-      .then((response) => {
-        if (fs.existsSync("./images/formIcons" + response.icon.replace("/img/formIcons", ""))) {
-          fs.unlink("./images/formIcons" + response.icon.replace("/img/formIcons", ""), (err) => {
-            if (err) console.log(err);
-          });
-        }
-        res.json(response);
-      })
-      .catch((err) => {
-        console.log(err);
-        res.status(404).send();
-      });
+      const formSetupData = {
+        formName: formStructure[1].formName,
+        description: formStructure[1].description,
+        formDetails: formStructure[0],
+        icon: `/img/formIcons/${req.file.filename}`,
+        primaryColor: formStructure[1].primaryColor,
+      };
+      formSetupModel
+        .findByIdAndUpdate({ _id: formStructure[2].form_id }, formSetupData)
+        .then((response) => {
+          if (fs.existsSync("./images/formIcons" + response.icon.replace("/img/formIcons", ""))) {
+            fs.unlink("./images/formIcons" + response.icon.replace("/img/formIcons", ""), (err) => {
+              if (err) console.log(err);
+            });
+          }
+          res.json(response);
+        })
+        .catch((err) => {
+          console.log(err);
+          res.status(404).send();
+        });
+    }
   } catch (error) {
     console.log(error);
     res.status(404).send();
   }
 });
 
-router.delete("/delete-forms", (req, res) => {
+const checkDeletionPermission = async (account, form_ids) => {
+  let isAllowed = true;
+  return new Promise((resolve, reject) => {
+    AccountModel.findById(account.account_id, { allowedForms: 1, _id: 0 })
+      .then((formResponse) => {
+        for (item of formResponse.allowedForms) {
+          //console.log(!form_ids.includes(item.formId) && item.permissionType !== "write");
+          if (!form_ids.includes(item.formId) || item.permissionType !== "write") {
+            isAllowed = false;
+            break;
+          }
+        }
+        resolve(isAllowed);
+      })
+      .catch((err) => {
+        console.log(error);
+      });
+  });
+};
+
+router.delete("/delete-forms", auth, async (req, res) => {
   try {
+    const form_ids = req.body["form_ids"];
+    let isAllowed = await checkDeletionPermission(req.account, form_ids);
     if (!mongoose.models.formSchemas) {
       createSchemasModel();
     }
-    const form_ids = req.body["form_ids"];
-
-    mongoose.models.formSchemas
-      .find({ _id: { $in: form_ids } }, { formName: 1 })
-      .then((response) => {
-        if (response) {
-          response.forEach((item) => {
-            mongoose.connection.dropCollection(item.formName, (err, result) => {});
-          });
+    if (req.account.role === "root" || (req.account.role === "admin" && isAllowed)) {
+      mongoose.models.formSchemas
+        .find({ _id: { $in: form_ids } }, { formName: 1 })
+        .then((response) => {
+          if (response) {
+            response.forEach((item) => {
+              mongoose.connection.dropCollection(item.formName, (err, result) => {});
+            });
+          }
+        })
+        .then(() => {
           mongoose.models.formSchemas
             .deleteMany({
               _id: { $in: form_ids },
@@ -131,11 +185,11 @@ router.delete("/delete-forms", (req, res) => {
             .catch((error) => {
               console.log(error);
             });
-        }
-      })
-      .catch((err) => {
-        console.log(err);
-      });
+        })
+        .catch((err) => {
+          console.log(err);
+        });
+    }
   } catch (error) {}
 });
 
@@ -173,7 +227,7 @@ router.post("/", (req, res) => {
   }
 });
 // Returns the requested form documents
-router.get("/", (req, res) => {
+router.get("/", verifyRootLevel, (req, res) => {
   try {
     if (!mongoose.models.formSchemas) {
       createSchemasModel();
@@ -207,29 +261,34 @@ router.get("/", (req, res) => {
 });
 
 // Update a document from a Form
-router.put("/", (req, res) => {
+router.put("/", auth, async (req, res) => {
   try {
-    if (!mongoose.models.formSchemas) {
-      createSchemasModel();
-    }
-    mongoose.models.formSchemas
-      .findOne({ _id: req.body.form_id })
-      .then((response) => {
-        if (response) {
-          const formModel = getModel(response);
-          formModel
-            .findByIdAndUpdate(req.body.document_id, req.body)
-            .then((response) => res.json(response))
-            .catch((err) => console.log(err));
+    const isAllowed = await checkPermission(req.account, req.body.form_id);
+    if (req.account.role === "root" || (req.account.role === "admin" && isAllowed)) {
+      if (!mongoose.models.formSchemas) {
+        createSchemasModel();
+      }
+      mongoose.models.formSchemas
+        .findOne({ _id: req.body.form_id })
+        .then((response) => {
+          if (response) {
+            const formModel = getModel(response);
+            formModel
+              .findByIdAndUpdate(req.body.document_id, req.body)
+              .then((response) => res.json(response))
+              .catch((err) => console.log(err));
+            return;
+          }
+          res.status(404).send("Form is unavailable");
           return;
-        }
-        res.status(404).send("Form is unavailable");
-        return;
-      })
-      .catch((err) => {
-        console.log(err);
-        res.status(404).send();
-      });
+        })
+        .catch((err) => {
+          console.log(err);
+          res.status(404).send();
+        });
+    } else {
+      res.status(401).send("You are not authorized for that operation");
+    }
   } catch (error) {
     console.log(error);
     res.status(404).send();
@@ -237,30 +296,35 @@ router.put("/", (req, res) => {
 });
 
 // Delete a Document from a form
-router.delete("/", (req, res) => {
+router.delete("/", auth, async (req, res) => {
   try {
-    const ids = req.body["document_ids"];
-    if (!mongoose.models.formSchemas) {
-      createSchemasModel();
-    }
-    mongoose.models.formSchemas
-      .findOne({ _id: req.body.form_id })
-      .then((response) => {
-        if (response) {
-          const formModel = getModel(response);
-          formModel
-            .deleteMany({ _id: { $in: ids } })
-            .then((response) => res.json(response))
-            .catch((err) => console.log(err));
+    const isAllowed = await checkPermission(req.account, req.body.form_id);
+    if (req.account.role === "root" || (req.account.role === "admin" && isAllowed)) {
+      const ids = req.body["document_ids"];
+      if (!mongoose.models.formSchemas) {
+        createSchemasModel();
+      }
+      mongoose.models.formSchemas
+        .findOne({ _id: req.body.form_id })
+        .then((response) => {
+          if (response) {
+            const formModel = getModel(response);
+            formModel
+              .deleteMany({ _id: { $in: ids } })
+              .then((response) => res.json(response))
+              .catch((err) => console.log(err));
+            return;
+          }
+          res.status(404).send("Form is unavailable");
           return;
-        }
-        res.status(404).send("Form is unavailable");
-        return;
-      })
-      .catch((err) => {
-        console.log(err);
-        res.status(404).send();
-      });
+        })
+        .catch((err) => {
+          console.log(err);
+          res.status(404).send();
+        });
+    }else{
+      res.status(401).send("You are not authorized for that operation.")
+    }
   } catch (error) {
     console.log(error);
     res.status(404).send();
@@ -285,6 +349,23 @@ const createSchemasModel = () => {
   };
   const formSetupSchema = new Schema(formUploadSetup);
   mongoose.model("formSchemas", formSetupSchema);
+};
+
+const checkPermission = async (account, form_id) => {
+  return new Promise((resolve, reject) => {
+    AccountModel.findById(account.account_id, { allowedForms: 1, _id: 0 })
+      .then((formResponse) => {
+        for (item of formResponse.allowedForms) {
+          if (item.formId === form_id && item.permissionType === "write") {
+            resolve(true);
+          }
+        }
+        resolve(false);
+      })
+      .catch((err) => {
+        reject(err);
+      });
+  });
 };
 
 module.exports = router;
